Extract test setup logic to CometFuzzTestBase
